ㅊㅐ영 심폐소생중

+++ src/main/java/com/ll/spirits/product/ProductRepository.java
@@ -1,6 +1,9 @@
 package com.ll.spirits.product;
 
+import com.ll.spirits.review.Review;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ProductRepository extends JpaRepository<Product, Long> {
+import java.util.List;
+
+public interface ProductRepository extends JpaRepository<Product, Integer> {
 }
